E2E test case
