+++ config/int/config.json
@@ -1,6 +1,7 @@
 {
 	"baseURL": "https://int-achieve-courseware-frontend.mldev.cloud/",
-	"loginURL": "https://int-achieve-iam.mldev.cloud/login?retURL=https://dev-achieve-courseware-frontend.mldev.cloud/courses",
+	"signURL":  "https://int-achieve-courseware-frontend.mldev.cloud/start",
+	"loginURL": "https://int-achieve-iam.mldev.cloud/login?retURL=https://int-achieve-courseware-frontend.mldev.cloud/courses",
 	"driver": {
 		"type": "chrome",
 		"headless": false
