+++ config/int/login/admin_alt.json
@@ -1,6 +1,6 @@
 {
-    "username": "ron.asher.contractor+ad01@macmillan.com",
-    "password": "Testing1234",
+    "username": "pgreeshma38@gmail.com",
+    "password": "Splended123$",
     "firstName": "Addy",
     "lastName": "min",
     "sq1": "What high school did you attend?",
