+++ config/int/login/instructor_1.json
@@ -1,6 +1,6 @@
 {
-    "username": "ron.asher.contractor+instr_01@macmillan.com",
-    "password": "Testing1234",
+    "username": "greeshma@bit-degree.com",
+    "password": "ABCabc@123",
     "firstName": "Instr",
     "lastName": "account",
     "sq1": "What high school did you attend?",
