+++ config/int/login/media_producer_1.json
@@ -1,6 +1,6 @@
 {
-    "username": "ron.asher.contractor+mp01@macmillan.com",
-    "password": "Testing1234",
+    "username": "greeshma.reddy.contractor@macmillan.com",
+    "password": "Splended123$",
     "firstName": "Media",
     "lastName": "Producer",
     "sq1": "What high school did you attend?",
